+++ pyiceberg/table/__init__.py
@@ -16,13 +16,10 @@
 # under the License.
 from __future__ import annotations
 
-import concurrent
 import itertools
 import uuid
 import warnings
 from abc import ABC, abstractmethod
-from collections import defaultdict
-from concurrent.futures import Future
 from dataclasses import dataclass
 from functools import cached_property
 from itertools import chain
@@ -31,7 +28,6 @@ from typing import (
     Any,
     Callable,
     Dict,
-    Generic,
     Iterable,
     List,
     Optional,
@@ -46,46 +42,32 @@ from sortedcontainers import SortedList
 
 import pyiceberg.expressions.parser as parser
 from pyiceberg.expressions import (
-    AlwaysFalse,
     AlwaysTrue,
     And,
     BooleanExpression,
     EqualTo,
-    Or,
-    Reference,
 )
 from pyiceberg.expressions.visitors import (
-    ROWS_MIGHT_NOT_MATCH,
-    ROWS_MUST_MATCH,
     _InclusiveMetricsEvaluator,
-    _StrictMetricsEvaluator,
     bind,
     expression_evaluator,
     inclusive_projection,
     manifest_evaluator,
 )
-from pyiceberg.io import FileIO, OutputFile, load_file_io
+from pyiceberg.io import FileIO, load_file_io
 from pyiceberg.manifest import (
     POSITIONAL_DELETE_SCHEMA,
     DataFile,
     DataFileContent,
     ManifestContent,
     ManifestEntry,
-    ManifestEntryStatus,
     ManifestFile,
-    ManifestWriter,
-    write_manifest,
-    write_manifest_list,
 )
 from pyiceberg.partitioning import (
-    INITIAL_PARTITION_SPEC_ID,
     PARTITION_FIELD_ID_START,
     UNPARTITIONED_PARTITION_SPEC,
-    PartitionField,
     PartitionKey,
     PartitionSpec,
-    _PartitionNameGenerator,
-    _visit_partition_field,
 )
 from pyiceberg.schema import Schema
 from pyiceberg.table.inspect import InspectTable
@@ -98,12 +80,8 @@ from pyiceberg.table.name_mapping import (
 )
 from pyiceberg.table.refs import SnapshotRef
 from pyiceberg.table.snapshots import (
-    Operation,
     Snapshot,
     SnapshotLogEntry,
-    SnapshotSummaryCollector,
-    Summary,
-    update_snapshot_summaries,
 )
 from pyiceberg.table.sorting import UNSORTED_SORT_ORDER, SortOrder
 from pyiceberg.table.update import (
@@ -112,7 +90,6 @@ from pyiceberg.table.update import (
     AddSnapshotUpdate,
     AddSortOrderUpdate,
     AssertCreate,
-    AssertLastAssignedPartitionId,
     AssertRefSnapshotId,
     AssertTableUUID,
     AssignUUIDUpdate,
@@ -125,14 +102,13 @@ from pyiceberg.table.update import (
     SetSnapshotRefUpdate,
     TableRequirement,
     TableUpdate,
-    U,
     UpdatesAndRequirements,
-    UpdateTableMetadata,
     UpgradeFormatVersionUpdate,
     update_table_metadata,
 )
 from pyiceberg.table.update.schema import UpdateSchema
-from pyiceberg.transforms import IdentityTransform, TimeTransform, Transform, VoidTransform
+from pyiceberg.table.update.snapshot import ManageSnapshots, UpdateSnapshot
+from pyiceberg.table.update.spec import UpdateSpec
 from pyiceberg.typedef import (
     EMPTY_DICT,
     IcebergBaseModel,
@@ -146,11 +122,10 @@ from pyiceberg.typedef import (
 from pyiceberg.types import (
     strtobool,
 )
-from pyiceberg.utils.bin_packing import ListPacker
 from pyiceberg.utils.concurrent import ExecutorFactory
 from pyiceberg.utils.config import Config
 from pyiceberg.utils.deprecated import deprecated, deprecation_message
-from pyiceberg.utils.properties import property_as_bool, property_as_int
+from pyiceberg.utils.properties import property_as_bool
 
 if TYPE_CHECKING:
     import daft
@@ -1452,84 +1427,6 @@ class DataScan(TableScan):
         return ray.data.from_arrow(self.to_arrow())
 
 
-class ManageSnapshots(UpdateTableMetadata["ManageSnapshots"]):
-    """
-    Run snapshot management operations using APIs.
-
-    APIs include create branch, create tag, etc.
-
-    Use table.manage_snapshots().<operation>().commit() to run a specific operation.
-    Use table.manage_snapshots().<operation-one>().<operation-two>().commit() to run multiple operations.
-    Pending changes are applied on commit.
-
-    We can also use context managers to make more changes. For example,
-
-    with table.manage_snapshots() as ms:
-       ms.create_tag(snapshot_id1, "Tag_A").create_tag(snapshot_id2, "Tag_B")
-    """
-
-    _updates: Tuple[TableUpdate, ...] = ()
-    _requirements: Tuple[TableRequirement, ...] = ()
-
-    def _commit(self) -> UpdatesAndRequirements:
-        """Apply the pending changes and commit."""
-        return self._updates, self._requirements
-
-    def create_tag(self, snapshot_id: int, tag_name: str, max_ref_age_ms: Optional[int] = None) -> ManageSnapshots:
-        """
-        Create a new tag pointing to the given snapshot id.
-
-        Args:
-            snapshot_id (int): snapshot id of the existing snapshot to tag
-            tag_name (str): name of the tag
-            max_ref_age_ms (Optional[int]): max ref age in milliseconds
-
-        Returns:
-            This for method chaining
-        """
-        update, requirement = self._transaction._set_ref_snapshot(
-            snapshot_id=snapshot_id,
-            ref_name=tag_name,
-            type="tag",
-            max_ref_age_ms=max_ref_age_ms,
-        )
-        self._updates += update
-        self._requirements += requirement
-        return self
-
-    def create_branch(
-        self,
-        snapshot_id: int,
-        branch_name: str,
-        max_ref_age_ms: Optional[int] = None,
-        max_snapshot_age_ms: Optional[int] = None,
-        min_snapshots_to_keep: Optional[int] = None,
-    ) -> ManageSnapshots:
-        """
-        Create a new branch pointing to the given snapshot id.
-
-        Args:
-            snapshot_id (int): snapshot id of existing snapshot at which the branch is created.
-            branch_name (str): name of the new branch
-            max_ref_age_ms (Optional[int]): max ref age in milliseconds
-            max_snapshot_age_ms (Optional[int]): max age of snapshots to keep in milliseconds
-            min_snapshots_to_keep (Optional[int]): min number of snapshots to keep in milliseconds
-        Returns:
-            This for method chaining
-        """
-        update, requirement = self._transaction._set_ref_snapshot(
-            snapshot_id=snapshot_id,
-            ref_name=branch_name,
-            type="branch",
-            max_ref_age_ms=max_ref_age_ms,
-            max_snapshot_age_ms=max_snapshot_age_ms,
-            min_snapshots_to_keep=min_snapshots_to_keep,
-        )
-        self._updates += update
-        self._requirements += requirement
-        return self
-
-
 @dataclass(frozen=True)
 class WriteTask:
     write_uuid: uuid.UUID
@@ -1558,16 +1455,6 @@ class AddFileTask:
     partition_field_value: Record
 
 
-def _new_manifest_path(location: str, num: int, commit_uuid: uuid.UUID) -> str:
-    return f"{location}/metadata/{commit_uuid}-m{num}.avro"
-
-
-def _generate_manifest_list_path(location: str, snapshot_id: int, attempt: int, commit_uuid: uuid.UUID) -> str:
-    # Mimics the behavior in Java:
-    # https://github.com/apache/iceberg/blob/c862b9177af8e2d83122220764a056f3b96fd00c/core/src/main/java/org/apache/iceberg/SnapshotProducer.java#L491
-    return f"{location}/metadata/snap-{snapshot_id}-{attempt}-{commit_uuid}.avro"
-
-
 def _parquet_files_to_data_files(table_metadata: TableMetadata, file_paths: List[str], io: FileIO) -> Iterable[DataFile]:
     """Convert a list files into DataFiles.
 
@@ -1577,869 +1464,3 @@ def _parquet_files_to_data_files(table_metadata: TableMetadata, file_paths: List
     from pyiceberg.io.pyarrow import parquet_files_to_data_files
 
     yield from parquet_files_to_data_files(io=io, table_metadata=table_metadata, file_paths=iter(file_paths))
-
-
-class _SnapshotProducer(UpdateTableMetadata[U], Generic[U]):
-    commit_uuid: uuid.UUID
-    _io: FileIO
-    _operation: Operation
-    _snapshot_id: int
-    _parent_snapshot_id: Optional[int]
-    _added_data_files: List[DataFile]
-    _manifest_num_counter: itertools.count[int]
-    _deleted_data_files: Set[DataFile]
-
-    def __init__(
-        self,
-        operation: Operation,
-        transaction: Transaction,
-        io: FileIO,
-        commit_uuid: Optional[uuid.UUID] = None,
-        snapshot_properties: Dict[str, str] = EMPTY_DICT,
-    ) -> None:
-        super().__init__(transaction)
-        self.commit_uuid = commit_uuid or uuid.uuid4()
-        self._io = io
-        self._operation = operation
-        self._snapshot_id = self._transaction.table_metadata.new_snapshot_id()
-        # Since we only support the main branch for now
-        self._parent_snapshot_id = (
-            snapshot.snapshot_id if (snapshot := self._transaction.table_metadata.current_snapshot()) else None
-        )
-        self._added_data_files = []
-        self._deleted_data_files = set()
-        self.snapshot_properties = snapshot_properties
-        self._manifest_num_counter = itertools.count(0)
-
-    def append_data_file(self, data_file: DataFile) -> _SnapshotProducer[U]:
-        self._added_data_files.append(data_file)
-        return self
-
-    def delete_data_file(self, data_file: DataFile) -> _SnapshotProducer[U]:
-        self._deleted_data_files.add(data_file)
-        return self
-
-    @abstractmethod
-    def _deleted_entries(self) -> List[ManifestEntry]: ...
-
-    @abstractmethod
-    def _existing_manifests(self) -> List[ManifestFile]: ...
-
-    def _process_manifests(self, manifests: List[ManifestFile]) -> List[ManifestFile]:
-        """To perform any post-processing on the manifests before writing them to the new snapshot."""
-        return manifests
-
-    def _manifests(self) -> List[ManifestFile]:
-        def _write_added_manifest() -> List[ManifestFile]:
-            if self._added_data_files:
-                with write_manifest(
-                    format_version=self._transaction.table_metadata.format_version,
-                    spec=self._transaction.table_metadata.spec(),
-                    schema=self._transaction.table_metadata.schema(),
-                    output_file=self.new_manifest_output(),
-                    snapshot_id=self._snapshot_id,
-                ) as writer:
-                    for data_file in self._added_data_files:
-                        writer.add(
-                            ManifestEntry(
-                                status=ManifestEntryStatus.ADDED,
-                                snapshot_id=self._snapshot_id,
-                                sequence_number=None,
-                                file_sequence_number=None,
-                                data_file=data_file,
-                            )
-                        )
-                return [writer.to_manifest_file()]
-            else:
-                return []
-
-        def _write_delete_manifest() -> List[ManifestFile]:
-            # Check if we need to mark the files as deleted
-            deleted_entries = self._deleted_entries()
-            if len(deleted_entries) > 0:
-                deleted_manifests = []
-                partition_groups: Dict[int, List[ManifestEntry]] = defaultdict(list)
-                for deleted_entry in deleted_entries:
-                    partition_groups[deleted_entry.data_file.spec_id].append(deleted_entry)
-                for spec_id, entries in partition_groups.items():
-                    with write_manifest(
-                        format_version=self._transaction.table_metadata.format_version,
-                        spec=self._transaction.table_metadata.specs()[spec_id],
-                        schema=self._transaction.table_metadata.schema(),
-                        output_file=self.new_manifest_output(),
-                        snapshot_id=self._snapshot_id,
-                    ) as writer:
-                        for entry in entries:
-                            writer.add_entry(entry)
-                    deleted_manifests.append(writer.to_manifest_file())
-                return deleted_manifests
-            else:
-                return []
-
-        executor = ExecutorFactory.get_or_create()
-
-        added_manifests = executor.submit(_write_added_manifest)
-        delete_manifests = executor.submit(_write_delete_manifest)
-        existing_manifests = executor.submit(self._existing_manifests)
-
-        return self._process_manifests(added_manifests.result() + delete_manifests.result() + existing_manifests.result())
-
-    def _summary(self, snapshot_properties: Dict[str, str] = EMPTY_DICT) -> Summary:
-        ssc = SnapshotSummaryCollector()
-        partition_summary_limit = int(
-            self._transaction.table_metadata.properties.get(
-                TableProperties.WRITE_PARTITION_SUMMARY_LIMIT, TableProperties.WRITE_PARTITION_SUMMARY_LIMIT_DEFAULT
-            )
-        )
-        ssc.set_partition_summary_limit(partition_summary_limit)
-
-        for data_file in self._added_data_files:
-            ssc.add_file(
-                data_file=data_file,
-                partition_spec=self._transaction.table_metadata.spec(),
-                schema=self._transaction.table_metadata.schema(),
-            )
-
-        if len(self._deleted_data_files) > 0:
-            specs = self._transaction.table_metadata.specs()
-            for data_file in self._deleted_data_files:
-                ssc.remove_file(
-                    data_file=data_file,
-                    partition_spec=specs[data_file.spec_id],
-                    schema=self._transaction.table_metadata.schema(),
-                )
-
-        previous_snapshot = (
-            self._transaction.table_metadata.snapshot_by_id(self._parent_snapshot_id)
-            if self._parent_snapshot_id is not None
-            else None
-        )
-
-        return update_snapshot_summaries(
-            summary=Summary(operation=self._operation, **ssc.build(), **snapshot_properties),
-            previous_summary=previous_snapshot.summary if previous_snapshot is not None else None,
-            truncate_full_table=self._operation == Operation.OVERWRITE,
-        )
-
-    def _commit(self) -> UpdatesAndRequirements:
-        new_manifests = self._manifests()
-        next_sequence_number = self._transaction.table_metadata.next_sequence_number()
-
-        summary = self._summary(self.snapshot_properties)
-
-        manifest_list_file_path = _generate_manifest_list_path(
-            location=self._transaction.table_metadata.location,
-            snapshot_id=self._snapshot_id,
-            attempt=0,
-            commit_uuid=self.commit_uuid,
-        )
-        with write_manifest_list(
-            format_version=self._transaction.table_metadata.format_version,
-            output_file=self._io.new_output(manifest_list_file_path),
-            snapshot_id=self._snapshot_id,
-            parent_snapshot_id=self._parent_snapshot_id,
-            sequence_number=next_sequence_number,
-        ) as writer:
-            writer.add_manifests(new_manifests)
-
-        snapshot = Snapshot(
-            snapshot_id=self._snapshot_id,
-            parent_snapshot_id=self._parent_snapshot_id,
-            manifest_list=manifest_list_file_path,
-            sequence_number=next_sequence_number,
-            summary=summary,
-            schema_id=self._transaction.table_metadata.current_schema_id,
-        )
-
-        return (
-            (
-                AddSnapshotUpdate(snapshot=snapshot),
-                SetSnapshotRefUpdate(
-                    snapshot_id=self._snapshot_id, parent_snapshot_id=self._parent_snapshot_id, ref_name="main", type="branch"
-                ),
-            ),
-            (AssertRefSnapshotId(snapshot_id=self._transaction.table_metadata.current_snapshot_id, ref="main"),),
-        )
-
-    @property
-    def snapshot_id(self) -> int:
-        return self._snapshot_id
-
-    def spec(self, spec_id: int) -> PartitionSpec:
-        return self._transaction.table_metadata.specs()[spec_id]
-
-    def new_manifest_writer(self, spec: PartitionSpec) -> ManifestWriter:
-        return write_manifest(
-            format_version=self._transaction.table_metadata.format_version,
-            spec=spec,
-            schema=self._transaction.table_metadata.schema(),
-            output_file=self.new_manifest_output(),
-            snapshot_id=self._snapshot_id,
-        )
-
-    def new_manifest_output(self) -> OutputFile:
-        return self._io.new_output(
-            _new_manifest_path(
-                location=self._transaction.table_metadata.location,
-                num=next(self._manifest_num_counter),
-                commit_uuid=self.commit_uuid,
-            )
-        )
-
-    def fetch_manifest_entry(self, manifest: ManifestFile, discard_deleted: bool = True) -> List[ManifestEntry]:
-        return manifest.fetch_manifest_entry(io=self._io, discard_deleted=discard_deleted)
-
-
-class DeleteFiles(_SnapshotProducer["DeleteFiles"]):
-    """Will delete manifest entries from the current snapshot based on the predicate.
-
-    This will produce a DELETE snapshot:
-        Data files were removed and their contents logically deleted and/or delete
-        files were added to delete rows.
-
-    From the specification
-    """
-
-    _predicate: BooleanExpression
-
-    def __init__(
-        self,
-        operation: Operation,
-        transaction: Transaction,
-        io: FileIO,
-        commit_uuid: Optional[uuid.UUID] = None,
-        snapshot_properties: Dict[str, str] = EMPTY_DICT,
-    ):
-        super().__init__(operation, transaction, io, commit_uuid, snapshot_properties)
-        self._predicate = AlwaysFalse()
-
-    def _commit(self) -> UpdatesAndRequirements:
-        # Only produce a commit when there is something to delete
-        if self.files_affected:
-            return super()._commit()
-        else:
-            return (), ()
-
-    def _build_partition_projection(self, spec_id: int) -> BooleanExpression:
-        schema = self._transaction.table_metadata.schema()
-        spec = self._transaction.table_metadata.specs()[spec_id]
-        project = inclusive_projection(schema, spec)
-        return project(self._predicate)
-
-    @cached_property
-    def partition_filters(self) -> KeyDefaultDict[int, BooleanExpression]:
-        return KeyDefaultDict(self._build_partition_projection)
-
-    def _build_manifest_evaluator(self, spec_id: int) -> Callable[[ManifestFile], bool]:
-        schema = self._transaction.table_metadata.schema()
-        spec = self._transaction.table_metadata.specs()[spec_id]
-        return manifest_evaluator(spec, schema, self.partition_filters[spec_id], case_sensitive=True)
-
-    def delete_by_predicate(self, predicate: BooleanExpression) -> None:
-        self._predicate = Or(self._predicate, predicate)
-
-    @cached_property
-    def _compute_deletes(self) -> Tuple[List[ManifestFile], List[ManifestEntry], bool]:
-        """Computes all the delete operation and cache it when nothing changes.
-
-        Returns:
-            - List of existing manifests that are not affected by the delete operation.
-            - The manifest-entries that are deleted based on the metadata.
-            - Flag indicating that rewrites of data-files are needed.
-        """
-        schema = self._transaction.table_metadata.schema()
-
-        def _copy_with_new_status(entry: ManifestEntry, status: ManifestEntryStatus) -> ManifestEntry:
-            return ManifestEntry(
-                status=status,
-                snapshot_id=entry.snapshot_id,
-                sequence_number=entry.sequence_number,
-                file_sequence_number=entry.file_sequence_number,
-                data_file=entry.data_file,
-            )
-
-        manifest_evaluators: Dict[int, Callable[[ManifestFile], bool]] = KeyDefaultDict(self._build_manifest_evaluator)
-        strict_metrics_evaluator = _StrictMetricsEvaluator(schema, self._predicate, case_sensitive=True).eval
-        inclusive_metrics_evaluator = _InclusiveMetricsEvaluator(schema, self._predicate, case_sensitive=True).eval
-
-        existing_manifests = []
-        total_deleted_entries = []
-        partial_rewrites_needed = False
-        self._deleted_data_files = set()
-        if snapshot := self._transaction.table_metadata.current_snapshot():
-            for manifest_file in snapshot.manifests(io=self._io):
-                if manifest_file.content == ManifestContent.DATA:
-                    if not manifest_evaluators[manifest_file.partition_spec_id](manifest_file):
-                        # If the manifest isn't relevant, we can just keep it in the manifest-list
-                        existing_manifests.append(manifest_file)
-                    else:
-                        # It is relevant, let's check out the content
-                        deleted_entries = []
-                        existing_entries = []
-                        for entry in manifest_file.fetch_manifest_entry(io=self._io, discard_deleted=True):
-                            if strict_metrics_evaluator(entry.data_file) == ROWS_MUST_MATCH:
-                                # Based on the metadata, it can be dropped right away
-                                deleted_entries.append(_copy_with_new_status(entry, ManifestEntryStatus.DELETED))
-                                self._deleted_data_files.add(entry.data_file)
-                            else:
-                                # Based on the metadata, we cannot determine if it can be deleted
-                                existing_entries.append(_copy_with_new_status(entry, ManifestEntryStatus.EXISTING))
-                                if inclusive_metrics_evaluator(entry.data_file) != ROWS_MIGHT_NOT_MATCH:
-                                    partial_rewrites_needed = True
-
-                        if len(deleted_entries) > 0:
-                            total_deleted_entries += deleted_entries
-
-                            # Rewrite the manifest
-                            if len(existing_entries) > 0:
-                                with write_manifest(
-                                    format_version=self._transaction.table_metadata.format_version,
-                                    spec=self._transaction.table_metadata.specs()[manifest_file.partition_spec_id],
-                                    schema=self._transaction.table_metadata.schema(),
-                                    output_file=self.new_manifest_output(),
-                                    snapshot_id=self._snapshot_id,
-                                ) as writer:
-                                    for existing_entry in existing_entries:
-                                        writer.add_entry(existing_entry)
-                                existing_manifests.append(writer.to_manifest_file())
-                        else:
-                            existing_manifests.append(manifest_file)
-                else:
-                    existing_manifests.append(manifest_file)
-
-        return existing_manifests, total_deleted_entries, partial_rewrites_needed
-
-    def _existing_manifests(self) -> List[ManifestFile]:
-        return self._compute_deletes[0]
-
-    def _deleted_entries(self) -> List[ManifestEntry]:
-        return self._compute_deletes[1]
-
-    @property
-    def rewrites_needed(self) -> bool:
-        """Indicate if data files need to be rewritten."""
-        return self._compute_deletes[2]
-
-    @property
-    def files_affected(self) -> bool:
-        """Indicate if any manifest-entries can be dropped."""
-        return len(self._deleted_entries()) > 0
-
-
-class FastAppendFiles(_SnapshotProducer["FastAppendFiles"]):
-    def _existing_manifests(self) -> List[ManifestFile]:
-        """To determine if there are any existing manifest files.
-
-        A fast append will add another ManifestFile to the ManifestList.
-        All the existing manifest files are considered existing.
-        """
-        existing_manifests = []
-
-        if self._parent_snapshot_id is not None:
-            previous_snapshot = self._transaction.table_metadata.snapshot_by_id(self._parent_snapshot_id)
-
-            if previous_snapshot is None:
-                raise ValueError(f"Snapshot could not be found: {self._parent_snapshot_id}")
-
-            for manifest in previous_snapshot.manifests(io=self._io):
-                if manifest.has_added_files() or manifest.has_existing_files() or manifest.added_snapshot_id == self._snapshot_id:
-                    existing_manifests.append(manifest)
-
-        return existing_manifests
-
-    def _deleted_entries(self) -> List[ManifestEntry]:
-        """To determine if we need to record any deleted manifest entries.
-
-        In case of an append, nothing is deleted.
-        """
-        return []
-
-
-class MergeAppendFiles(FastAppendFiles):
-    _target_size_bytes: int
-    _min_count_to_merge: int
-    _merge_enabled: bool
-
-    def __init__(
-        self,
-        operation: Operation,
-        transaction: Transaction,
-        io: FileIO,
-        commit_uuid: Optional[uuid.UUID] = None,
-        snapshot_properties: Dict[str, str] = EMPTY_DICT,
-    ) -> None:
-        super().__init__(operation, transaction, io, commit_uuid, snapshot_properties)
-        self._target_size_bytes = property_as_int(
-            self._transaction.table_metadata.properties,
-            TableProperties.MANIFEST_TARGET_SIZE_BYTES,
-            TableProperties.MANIFEST_TARGET_SIZE_BYTES_DEFAULT,
-        )  # type: ignore
-        self._min_count_to_merge = property_as_int(
-            self._transaction.table_metadata.properties,
-            TableProperties.MANIFEST_MIN_MERGE_COUNT,
-            TableProperties.MANIFEST_MIN_MERGE_COUNT_DEFAULT,
-        )  # type: ignore
-        self._merge_enabled = property_as_bool(
-            self._transaction.table_metadata.properties,
-            TableProperties.MANIFEST_MERGE_ENABLED,
-            TableProperties.MANIFEST_MERGE_ENABLED_DEFAULT,
-        )
-
-    def _process_manifests(self, manifests: List[ManifestFile]) -> List[ManifestFile]:
-        """To perform any post-processing on the manifests before writing them to the new snapshot.
-
-        In MergeAppendFiles, we merge manifests based on the target size and the minimum count to merge
-        if automatic merge is enabled.
-        """
-        unmerged_data_manifests = [manifest for manifest in manifests if manifest.content == ManifestContent.DATA]
-        unmerged_deletes_manifests = [manifest for manifest in manifests if manifest.content == ManifestContent.DELETES]
-
-        data_manifest_merge_manager = _ManifestMergeManager(
-            target_size_bytes=self._target_size_bytes,
-            min_count_to_merge=self._min_count_to_merge,
-            merge_enabled=self._merge_enabled,
-            snapshot_producer=self,
-        )
-
-        return data_manifest_merge_manager.merge_manifests(unmerged_data_manifests) + unmerged_deletes_manifests
-
-
-class OverwriteFiles(_SnapshotProducer["OverwriteFiles"]):
-    """Overwrites data from the table. This will produce an OVERWRITE snapshot.
-
-    Data and delete files were added and removed in a logical overwrite operation.
-    """
-
-    def _existing_manifests(self) -> List[ManifestFile]:
-        """Determine if there are any existing manifest files."""
-        existing_files = []
-
-        if snapshot := self._transaction.table_metadata.current_snapshot():
-            for manifest_file in snapshot.manifests(io=self._io):
-                entries = manifest_file.fetch_manifest_entry(io=self._io, discard_deleted=True)
-                found_deleted_data_files = [entry.data_file for entry in entries if entry.data_file in self._deleted_data_files]
-
-                if len(found_deleted_data_files) == 0:
-                    existing_files.append(manifest_file)
-                else:
-                    # We have to rewrite the manifest file without the deleted data files
-                    if any(entry.data_file not in found_deleted_data_files for entry in entries):
-                        with write_manifest(
-                            format_version=self._transaction.table_metadata.format_version,
-                            spec=self._transaction.table_metadata.spec(),
-                            schema=self._transaction.table_metadata.schema(),
-                            output_file=self.new_manifest_output(),
-                            snapshot_id=self._snapshot_id,
-                        ) as writer:
-                            [
-                                writer.add_entry(
-                                    ManifestEntry(
-                                        status=ManifestEntryStatus.EXISTING,
-                                        snapshot_id=entry.snapshot_id,
-                                        sequence_number=entry.sequence_number,
-                                        file_sequence_number=entry.file_sequence_number,
-                                        data_file=entry.data_file,
-                                    )
-                                )
-                                for entry in entries
-                                if entry.data_file not in found_deleted_data_files
-                            ]
-                        existing_files.append(writer.to_manifest_file())
-        return existing_files
-
-    def _deleted_entries(self) -> List[ManifestEntry]:
-        """To determine if we need to record any deleted entries.
-
-        With a full overwrite all the entries are considered deleted.
-        With partial overwrites we have to use the predicate to evaluate
-        which entries are affected.
-        """
-        if self._parent_snapshot_id is not None:
-            previous_snapshot = self._transaction.table_metadata.snapshot_by_id(self._parent_snapshot_id)
-            if previous_snapshot is None:
-                # This should never happen since you cannot overwrite an empty table
-                raise ValueError(f"Could not find the previous snapshot: {self._parent_snapshot_id}")
-
-            executor = ExecutorFactory.get_or_create()
-
-            def _get_entries(manifest: ManifestFile) -> List[ManifestEntry]:
-                return [
-                    ManifestEntry(
-                        status=ManifestEntryStatus.DELETED,
-                        snapshot_id=entry.snapshot_id,
-                        sequence_number=entry.sequence_number,
-                        file_sequence_number=entry.file_sequence_number,
-                        data_file=entry.data_file,
-                    )
-                    for entry in manifest.fetch_manifest_entry(self._io, discard_deleted=True)
-                    if entry.data_file.content == DataFileContent.DATA and entry.data_file in self._deleted_data_files
-                ]
-
-            list_of_entries = executor.map(_get_entries, previous_snapshot.manifests(self._io))
-            return list(chain(*list_of_entries))
-        else:
-            return []
-
-
-class UpdateSnapshot:
-    _transaction: Transaction
-    _io: FileIO
-    _snapshot_properties: Dict[str, str]
-
-    def __init__(self, transaction: Transaction, io: FileIO, snapshot_properties: Dict[str, str] = EMPTY_DICT) -> None:
-        self._transaction = transaction
-        self._io = io
-        self._snapshot_properties = snapshot_properties
-
-    def fast_append(self) -> FastAppendFiles:
-        return FastAppendFiles(
-            operation=Operation.APPEND, transaction=self._transaction, io=self._io, snapshot_properties=self._snapshot_properties
-        )
-
-    def merge_append(self) -> MergeAppendFiles:
-        return MergeAppendFiles(
-            operation=Operation.APPEND, transaction=self._transaction, io=self._io, snapshot_properties=self._snapshot_properties
-        )
-
-    def overwrite(self, commit_uuid: Optional[uuid.UUID] = None) -> OverwriteFiles:
-        return OverwriteFiles(
-            commit_uuid=commit_uuid,
-            operation=Operation.OVERWRITE
-            if self._transaction.table_metadata.current_snapshot() is not None
-            else Operation.APPEND,
-            transaction=self._transaction,
-            io=self._io,
-            snapshot_properties=self._snapshot_properties,
-        )
-
-    def delete(self) -> DeleteFiles:
-        return DeleteFiles(
-            operation=Operation.DELETE,
-            transaction=self._transaction,
-            io=self._io,
-            snapshot_properties=self._snapshot_properties,
-        )
-
-
-class UpdateSpec(UpdateTableMetadata["UpdateSpec"]):
-    _transaction: Transaction
-    _name_to_field: Dict[str, PartitionField] = {}
-    _name_to_added_field: Dict[str, PartitionField] = {}
-    _transform_to_field: Dict[Tuple[int, str], PartitionField] = {}
-    _transform_to_added_field: Dict[Tuple[int, str], PartitionField] = {}
-    _renames: Dict[str, str] = {}
-    _added_time_fields: Dict[int, PartitionField] = {}
-    _case_sensitive: bool
-    _adds: List[PartitionField]
-    _deletes: Set[int]
-    _last_assigned_partition_id: int
-
-    def __init__(self, transaction: Transaction, case_sensitive: bool = True) -> None:
-        super().__init__(transaction)
-        self._name_to_field = {field.name: field for field in transaction.table_metadata.spec().fields}
-        self._name_to_added_field = {}
-        self._transform_to_field = {
-            (field.source_id, repr(field.transform)): field for field in transaction.table_metadata.spec().fields
-        }
-        self._transform_to_added_field = {}
-        self._adds = []
-        self._deletes = set()
-        self._last_assigned_partition_id = transaction.table_metadata.last_partition_id or PARTITION_FIELD_ID_START - 1
-        self._renames = {}
-        self._transaction = transaction
-        self._case_sensitive = case_sensitive
-        self._added_time_fields = {}
-
-    def add_field(
-        self,
-        source_column_name: str,
-        transform: Transform[Any, Any],
-        partition_field_name: Optional[str] = None,
-    ) -> UpdateSpec:
-        ref = Reference(source_column_name)
-        bound_ref = ref.bind(self._transaction.table_metadata.schema(), self._case_sensitive)
-        # verify transform can actually bind it
-        output_type = bound_ref.field.field_type
-        if not transform.can_transform(output_type):
-            raise ValueError(f"{transform} cannot transform {output_type} values from {bound_ref.field.name}")
-
-        transform_key = (bound_ref.field.field_id, repr(transform))
-        existing_partition_field = self._transform_to_field.get(transform_key)
-        if existing_partition_field and self._is_duplicate_partition(transform, existing_partition_field):
-            raise ValueError(f"Duplicate partition field for ${ref.name}=${ref}, ${existing_partition_field} already exists")
-
-        added = self._transform_to_added_field.get(transform_key)
-        if added:
-            raise ValueError(f"Already added partition: {added.name}")
-
-        new_field = self._partition_field((bound_ref.field.field_id, transform), partition_field_name)
-        if new_field.name in self._name_to_added_field:
-            raise ValueError(f"Already added partition field with name: {new_field.name}")
-
-        if isinstance(new_field.transform, TimeTransform):
-            existing_time_field = self._added_time_fields.get(new_field.source_id)
-            if existing_time_field:
-                raise ValueError(f"Cannot add time partition field: {new_field.name} conflicts with {existing_time_field.name}")
-            self._added_time_fields[new_field.source_id] = new_field
-        self._transform_to_added_field[transform_key] = new_field
-
-        existing_partition_field = self._name_to_field.get(new_field.name)
-        if existing_partition_field and new_field.field_id not in self._deletes:
-            if isinstance(existing_partition_field.transform, VoidTransform):
-                self.rename_field(
-                    existing_partition_field.name, existing_partition_field.name + "_" + str(existing_partition_field.field_id)
-                )
-            else:
-                raise ValueError(f"Cannot add duplicate partition field name: {existing_partition_field.name}")
-
-        self._name_to_added_field[new_field.name] = new_field
-        self._adds.append(new_field)
-        return self
-
-    def add_identity(self, source_column_name: str) -> UpdateSpec:
-        return self.add_field(source_column_name, IdentityTransform(), None)
-
-    def remove_field(self, name: str) -> UpdateSpec:
-        added = self._name_to_added_field.get(name)
-        if added:
-            raise ValueError(f"Cannot delete newly added field {name}")
-        renamed = self._renames.get(name)
-        if renamed:
-            raise ValueError(f"Cannot rename and delete field {name}")
-        field = self._name_to_field.get(name)
-        if not field:
-            raise ValueError(f"No such partition field: {name}")
-
-        self._deletes.add(field.field_id)
-        return self
-
-    def rename_field(self, name: str, new_name: str) -> UpdateSpec:
-        existing_field = self._name_to_field.get(new_name)
-        if existing_field and isinstance(existing_field.transform, VoidTransform):
-            return self.rename_field(name, name + "_" + str(existing_field.field_id))
-        added = self._name_to_added_field.get(name)
-        if added:
-            raise ValueError("Cannot rename recently added partitions")
-        field = self._name_to_field.get(name)
-        if not field:
-            raise ValueError(f"Cannot find partition field {name}")
-        if field.field_id in self._deletes:
-            raise ValueError(f"Cannot delete and rename partition field {name}")
-        self._renames[name] = new_name
-        return self
-
-    def _commit(self) -> UpdatesAndRequirements:
-        new_spec = self._apply()
-        updates: Tuple[TableUpdate, ...] = ()
-        requirements: Tuple[TableRequirement, ...] = ()
-
-        if self._transaction.table_metadata.default_spec_id != new_spec.spec_id:
-            if new_spec.spec_id not in self._transaction.table_metadata.specs():
-                updates = (
-                    AddPartitionSpecUpdate(spec=new_spec),
-                    SetDefaultSpecUpdate(spec_id=-1),
-                )
-            else:
-                updates = (SetDefaultSpecUpdate(spec_id=new_spec.spec_id),)
-
-            required_last_assigned_partitioned_id = self._transaction.table_metadata.last_partition_id
-            requirements = (AssertLastAssignedPartitionId(last_assigned_partition_id=required_last_assigned_partitioned_id),)
-
-        return updates, requirements
-
-    def _apply(self) -> PartitionSpec:
-        def _check_and_add_partition_name(schema: Schema, name: str, source_id: int, partition_names: Set[str]) -> None:
-            try:
-                field = schema.find_field(name)
-            except ValueError:
-                field = None
-
-            if source_id is not None and field is not None and field.field_id != source_id:
-                raise ValueError(f"Cannot create identity partition from a different field in the schema {name}")
-            elif field is not None and source_id != field.field_id:
-                raise ValueError(f"Cannot create partition from name that exists in schema {name}")
-            if not name:
-                raise ValueError("Undefined name")
-            if name in partition_names:
-                raise ValueError(f"Partition name has to be unique: {name}")
-            partition_names.add(name)
-
-        def _add_new_field(
-            schema: Schema, source_id: int, field_id: int, name: str, transform: Transform[Any, Any], partition_names: Set[str]
-        ) -> PartitionField:
-            _check_and_add_partition_name(schema, name, source_id, partition_names)
-            return PartitionField(source_id, field_id, transform, name)
-
-        partition_fields = []
-        partition_names: Set[str] = set()
-        for field in self._transaction.table_metadata.spec().fields:
-            if field.field_id not in self._deletes:
-                renamed = self._renames.get(field.name)
-                if renamed:
-                    new_field = _add_new_field(
-                        self._transaction.table_metadata.schema(),
-                        field.source_id,
-                        field.field_id,
-                        renamed,
-                        field.transform,
-                        partition_names,
-                    )
-                else:
-                    new_field = _add_new_field(
-                        self._transaction.table_metadata.schema(),
-                        field.source_id,
-                        field.field_id,
-                        field.name,
-                        field.transform,
-                        partition_names,
-                    )
-                partition_fields.append(new_field)
-            elif self._transaction.table_metadata.format_version == 1:
-                renamed = self._renames.get(field.name)
-                if renamed:
-                    new_field = _add_new_field(
-                        self._transaction.table_metadata.schema(),
-                        field.source_id,
-                        field.field_id,
-                        renamed,
-                        VoidTransform(),
-                        partition_names,
-                    )
-                else:
-                    new_field = _add_new_field(
-                        self._transaction.table_metadata.schema(),
-                        field.source_id,
-                        field.field_id,
-                        field.name,
-                        VoidTransform(),
-                        partition_names,
-                    )
-
-                partition_fields.append(new_field)
-
-        for added_field in self._adds:
-            new_field = PartitionField(
-                source_id=added_field.source_id,
-                field_id=added_field.field_id,
-                transform=added_field.transform,
-                name=added_field.name,
-            )
-            partition_fields.append(new_field)
-
-        # Reuse spec id or create a new one.
-        new_spec = PartitionSpec(*partition_fields)
-        new_spec_id = INITIAL_PARTITION_SPEC_ID
-        for spec in self._transaction.table_metadata.specs().values():
-            if new_spec.compatible_with(spec):
-                new_spec_id = spec.spec_id
-                break
-            elif new_spec_id <= spec.spec_id:
-                new_spec_id = spec.spec_id + 1
-        return PartitionSpec(*partition_fields, spec_id=new_spec_id)
-
-    def _partition_field(self, transform_key: Tuple[int, Transform[Any, Any]], name: Optional[str]) -> PartitionField:
-        if self._transaction.table_metadata.format_version == 2:
-            source_id, transform = transform_key
-            historical_fields = []
-            for spec in self._transaction.table_metadata.specs().values():
-                for field in spec.fields:
-                    historical_fields.append((field.source_id, field.field_id, repr(field.transform), field.name))
-
-            for field_key in historical_fields:
-                if field_key[0] == source_id and field_key[2] == repr(transform):
-                    if name is None or field_key[3] == name:
-                        return PartitionField(source_id, field_key[1], transform, name)
-
-        new_field_id = self._new_field_id()
-        if name is None:
-            tmp_field = PartitionField(transform_key[0], new_field_id, transform_key[1], "unassigned_field_name")
-            name = _visit_partition_field(self._transaction.table_metadata.schema(), tmp_field, _PartitionNameGenerator())
-        return PartitionField(transform_key[0], new_field_id, transform_key[1], name)
-
-    def _new_field_id(self) -> int:
-        self._last_assigned_partition_id += 1
-        return self._last_assigned_partition_id
-
-    def _is_duplicate_partition(self, transform: Transform[Any, Any], partition_field: PartitionField) -> bool:
-        return partition_field.field_id not in self._deletes and partition_field.transform == transform
-
-
-class _ManifestMergeManager(Generic[U]):
-    _target_size_bytes: int
-    _min_count_to_merge: int
-    _merge_enabled: bool
-    _snapshot_producer: _SnapshotProducer[U]
-
-    def __init__(
-        self, target_size_bytes: int, min_count_to_merge: int, merge_enabled: bool, snapshot_producer: _SnapshotProducer[U]
-    ) -> None:
-        self._target_size_bytes = target_size_bytes
-        self._min_count_to_merge = min_count_to_merge
-        self._merge_enabled = merge_enabled
-        self._snapshot_producer = snapshot_producer
-
-    def _group_by_spec(self, manifests: List[ManifestFile]) -> Dict[int, List[ManifestFile]]:
-        groups = defaultdict(list)
-        for manifest in manifests:
-            groups[manifest.partition_spec_id].append(manifest)
-        return groups
-
-    def _create_manifest(self, spec_id: int, manifest_bin: List[ManifestFile]) -> ManifestFile:
-        with self._snapshot_producer.new_manifest_writer(spec=self._snapshot_producer.spec(spec_id)) as writer:
-            for manifest in manifest_bin:
-                for entry in self._snapshot_producer.fetch_manifest_entry(manifest=manifest, discard_deleted=False):
-                    if entry.status == ManifestEntryStatus.DELETED and entry.snapshot_id == self._snapshot_producer.snapshot_id:
-                        #  only files deleted by this snapshot should be added to the new manifest
-                        writer.delete(entry)
-                    elif entry.status == ManifestEntryStatus.ADDED and entry.snapshot_id == self._snapshot_producer.snapshot_id:
-                        # added entries from this snapshot are still added, otherwise they should be existing
-                        writer.add(entry)
-                    elif entry.status != ManifestEntryStatus.DELETED:
-                        # add all non-deleted files from the old manifest as existing files
-                        writer.existing(entry)
-
-        return writer.to_manifest_file()
-
-    def _merge_group(self, first_manifest: ManifestFile, spec_id: int, manifests: List[ManifestFile]) -> List[ManifestFile]:
-        packer: ListPacker[ManifestFile] = ListPacker(target_weight=self._target_size_bytes, lookback=1, largest_bin_first=False)
-        bins: List[List[ManifestFile]] = packer.pack_end(manifests, lambda m: m.manifest_length)
-
-        def merge_bin(manifest_bin: List[ManifestFile]) -> List[ManifestFile]:
-            output_manifests = []
-            if len(manifest_bin) == 1:
-                output_manifests.append(manifest_bin[0])
-            elif first_manifest in manifest_bin and len(manifest_bin) < self._min_count_to_merge:
-                #  if the bin has the first manifest (the new data files or an appended manifest file) then only
-                #  merge it if the number of manifests is above the minimum count. this is applied only to bins
-                #  with an in-memory manifest so that large manifests don't prevent merging older groups.
-                output_manifests.extend(manifest_bin)
-            else:
-                output_manifests.append(self._create_manifest(spec_id, manifest_bin))
-
-            return output_manifests
-
-        executor = ExecutorFactory.get_or_create()
-        futures = [executor.submit(merge_bin, b) for b in bins]
-
-        # for consistent ordering, we need to maintain future order
-        futures_index = {f: i for i, f in enumerate(futures)}
-        completed_futures: SortedList[Future[List[ManifestFile]]] = SortedList(iterable=[], key=lambda f: futures_index[f])
-        for future in concurrent.futures.as_completed(futures):
-            completed_futures.add(future)
-
-        bin_results: List[List[ManifestFile]] = [f.result() for f in completed_futures if f.result()]
-
-        return [manifest for bin_result in bin_results for manifest in bin_result]
-
-    def merge_manifests(self, manifests: List[ManifestFile]) -> List[ManifestFile]:
-        if not self._merge_enabled or len(manifests) == 0:
-            return manifests
-
-        first_manifest = manifests[0]
-        groups = self._group_by_spec(manifests)
-
-        merged_manifests = []
-        for spec_id in reversed(groups.keys()):
-            merged_manifests.extend(self._merge_group(first_manifest, spec_id, groups[spec_id]))
-
-        return merged_manifests
