+++ tests/integration/test_rest_schema.py
@@ -22,9 +22,10 @@ from pyiceberg.catalog import Catalog, load_catalog
 from pyiceberg.exceptions import CommitFailedException, NoSuchTableError, ValidationError
 from pyiceberg.partitioning import PartitionField, PartitionSpec
 from pyiceberg.schema import Schema, prune_columns
-from pyiceberg.table import Table, TableProperties, UpdateSchema
+from pyiceberg.table import Table, TableProperties
 from pyiceberg.table.name_mapping import MappedField, NameMapping, create_mapping_from_schema
 from pyiceberg.table.sorting import SortField, SortOrder
+from pyiceberg.table.update.schema import UpdateSchema
 from pyiceberg.transforms import IdentityTransform
 from pyiceberg.types import (
     BinaryType,
